// src/app/Models/Product.ts
// products.ts (or any appropriate file name)
export interface Product {
    name: string;
    price: number;
    yearOfManufacture: string;
    manufacturer: string;
    condition: string;
    category: string;
    color: string;
    description: string;
    image: File;
    // Add other properties as needed
  }
  